refactor: further split controller
